types update refactoring
